#1: Fundamental Iro support
This currently supports as much of the syntax as possible, but only for
one or two morphemes from each set. For instance, it currently only
supports double quotes for regular strings and only . subscription into
keys for references.

Later, this will be converted to a Jinja2 template which generates the
various string forms, so as to not make the source itself too unwieldy.